Centralise filterModel_Label method
