+++ src/components/labels.js
@@ -36,4 +36,34 @@ function buildLabels() {
 	})
 }
 
-export { buildLabels }
+function filterModel_Label(model) {
+	// need to match them all so set up an array of matches
+	let result = true
+	Object.values(this.label).forEach(v => {
+		let labelFilter = this.filterValue('value', v)
+		if (!labelFilter) return
+
+		// set up a bool for if the label had been found
+		let labelFound = false
+		model._original.labels.forEach(label => {
+			if (label.id == v.id) {
+				// if this is the right label mark as found
+				labelFound = true
+				// now check for a match - update result if false
+				let filterResult = false
+				labelFilter.forEach(filter => {
+					// if we find at least one match then set this filterResult to true
+					if (label.value.includes(filter)) filterResult = true
+				})
+
+				if (!filterResult) result = false
+			}
+		})
+		// if this group doesn't have this label then update result
+		if (!labelFound) result = false
+	})
+
+	return result
+}
+
+export { buildLabels, filterModel_Label }
