+++ src/smallgroups/CSGroups.js
@@ -1,6 +1,6 @@
 import Base from '../base'
 import Group from './group'
-import { buildLabels } from '../components/labels'
+import { buildLabels, filterModel_Label } from '../components/labels'
 
 export default class CSGroups extends Base {
 	buildModelObject = function (model) {
@@ -60,36 +60,6 @@ export default class CSGroups extends Base {
 		if (model.day == null) return true
 		// return on id or name for legacy support
 		return dayValue.includes(model.day.format('dddd')) || dayValue.includes(model._original.day)
-	}
-
-	filterModel_Label = function (model) {
-		// need to match them all so set up an array of matches
-		let result = true
-		Object.values(this.label).forEach(v => {
-			let labelFilter = this.filterValue('value', v)
-			if (!labelFilter) return
-
-			// set up a bool for if the label had been found
-			let labelFound = false
-			model._original.labels.forEach(label => {
-				if (label.id == v.id) {
-					// if this is the right label mark as found
-					labelFound = true
-					// now check for a match - update result if false
-					let filterResult = false
-					labelFilter.forEach(filter => {
-						// if we find at least one match then set this filterResult to true
-						if (label.value.includes(filter)) filterResult = true
-					})
-
-					if (!filterResult) result = false
-				}
-			})
-			// if this group doesn't have this label then update result
-			if (!labelFound) result = false
-		})
-
-		return result
 	}
 
 	filterModel_Search = function (model) {
@@ -207,5 +177,6 @@ export default class CSGroups extends Base {
 
 		// shared function between label-using classes
 		this.buildLabels = buildLabels
+		this.filterModel_Label = filterModel_Label
 	}
 }
